move shorteting logic to ShortenHelper.cs
